requirements.txt
absl-py==2.0.0
accelerate==1.0.1
aiohttp==3.8.6
aiosignal==1.3.1
airportsdata==20241001
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argilla==1.16.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.3
bitarray==3.0.0
black==21.4b2
blinker==1.8.2
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
click==8.1.7
cloudpickle==2.2.1
colbert-ai==0.2.19
coloredlogs==15.0.1
colpali_engine==0.3.2
comm==0.1.4
commonmark==0.9.1
contourpy==1.1.1
cozo_embedded==0.7.5
cryptography==41.0.4
cycler==0.12.1
dataclasses-json==0.6.1
datasets==3.0.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@d1e04565d3bec8719335b88be9e9b961bf3ec464
dill==0.3.8
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
docx2txt==0.8
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
faiss-cpu==1.9.0
fast-pytorch-kmeans==0.2.0.1
fastapi==0.103.2
fastapi-cli==0.0.5
filelock==3.12.4
filetype==1.2.0
Flask==3.0.3
flatbuffers==23.5.26
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.9.2
future==0.18.3
fvcore==0.1.5.post20221221
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.3
google-auth-oauthlib==1.0.0
GPUtil==1.4.0
graphviz==0.20.1
greenlet==3.1.1
grpcio==1.59.0
h11==0.14.0
hf_transfer==0.1.8
httpcore==0.16.3
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.25.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
importlib-resources==6.1.0
interegular==0.3.3
iopath==0.1.9
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.6.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsons==1.6.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.3.1
jupyter_core==5.4.0
kiwisolver==1.4.5
langchain==0.3.4
langchain-core==0.3.12
langchain-text-splitters==0.3.0
langdetect==1.0.9
langsmith==0.1.136
lark==1.2.2
layoutparser==0.3.4
llama-cloud==0.1.4
llama-index==0.11.19
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.19
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-openai==0.2.15
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.10
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mlx==0.18.1
mlx-lm==0.19.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
ninja==1.11.1.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.15.1
openai==1.52.0
opencv-python==4.8.1.78
openpyxl==3.1.2
orjson==3.10.9
outcome==1.3.0.post0
outlines==0.1.0
outlines_core==0.1.0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
parso==0.8.3
pathspec==0.11.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
peft==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
portalocker==2.8.2
posthog==3.0.2
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.5
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycountry==24.6.1
pycozo==0.7.5
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
pydot==1.4.2
Pygments==2.16.1
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pynvml==11.5.3
pyodata==1.10.1
pyodbc==5.0.1
pyparsing==3.1.1
pypdf==4.3.1
pypdfium2==4.20.0
PyPika==0.48.9
PySocks==1.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.22
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
RAGatouille==0.0.8.post4
rapidfuzz==3.4.0
referencing==0.35.1
regex==2023.10.3
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.11.3
selenium==4.24.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.21
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
striprtf==0.0.26
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
termcolor==2.3.0
threadpoolctl==3.5.0
tiktoken==0.5.1
timm==0.9.7
tokenizers==0.20.1
toml==0.10.2
torch==2.5.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.5
traitlets==5.11.2
transformers @ git+https://github.com/huggingface/transformers@c9afee539204f5e658d03e63a1df3aacb4cab305
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
typish==1.9.3
ujson==5.10.0
undetected-chromedriver==3.5.5
unstructured==0.10.19
unstructured-inference==0.7.2
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
voyager==2.0.9
Wand==0.6.11
watchfiles==0.20.0
wcwidth==0.2.8
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.0
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.7
xxhash==3.5.0
yacs==0.1.8
yarl==1.9.2
